adit_server/pyproject.toml
[project]
name = "adit_server"
version = "0.0.0"
requires-python = ">=3.12,<4.0"
dependencies = [
    "adit-radis-shared @ git+https://github.com/openradx/adit-radis-shared.git@0.14.0",
    "adrf",
    "aiofiles",
    "asyncinotify",
    "channels",
    "crispy-bootstrap5",
    "cryptography",
    "daphne",
    "dicognito<0.18.0",
    "dicomweb-client",
    "Django",
    "django-crispy-forms",
    "django-dbbackup",
    "django-extensions",
    "django-filter",
    "django-htmx",
    "django-loginas",
    "django-pglock",
    "django-registration-redux",
    "django-revproxy",
    "django-tables2",
    "djangorestframework",
    "environs[django]",
    "humanize",
    "janus",
    "Markdown",
    "openpyxl",
    "pandas",
    "pebble",
    "procrastinate[django]",
    "psycopg[binary]",
    "pyarrow",
    "pydicom<3.0.0",
    "pynetdicom",
    "Twisted[tls,http2]",
    "wait-for-it",
    "watchfiles",
    "whitenoise",
]
